requirements.txt
Django==5.2.4
djangorestframework==3.16.0
drf-spectacular==0.28.0
numpy==2.3.1
pandas==2.3.1
requests==2.32.4
